add confirmation on deletion of data sources (#19)

+++ console/src/components/Services/Data/Schema/ManageDatabase.tsx
@@ -16,6 +16,7 @@ import {
   addDataSource,
 } from '../../../../metadata/actions';
 import { RightContainer } from '../../../Common/Layout/RightContainer';
+import { getConfirmation } from '../../../Common/utils/jsUtils';
 
 const dummyData: DataSource[] = [
   {
@@ -97,7 +98,14 @@ const ManageDatabase: React.FC<ManageDatabaseInjectedProps> = ({
   dispatch,
 }) => {
   const onRemove = (name: string, driver: Driver, cb: () => void) => {
-    (dispatch(removeDataSource({ driver, name })) as any).then(cb); // todo
+    const confirmation = getConfirmation(
+      `Your action will remove the "${name}" data source`,
+      true,
+      name
+    );
+    if (confirmation) {
+      (dispatch(removeDataSource({ driver, name })) as any).then(cb); // todo
+    }
   };
 
   const onReload = (name: string, driver: Driver, cb: () => void) => {
@@ -143,13 +151,15 @@ const ManageDatabase: React.FC<ManageDatabaseInjectedProps> = ({
   return (
     <RightContainer>
       <div
-        className={`container-fluid ${styles.padd_left_remove} ${styles.padd_top} ${styles.manage_dbs_page}`}
+        className={`container-fluid ${styles.padd_left_remove} ${styles.manage_dbs_page}`}
       >
         <div className={styles.padd_left}>
           <Helmet title="Manage - Data | Hasura" />
           <BreadCrumb breadCrumbs={crumbs} />
           <div className={styles.display_flex}>
-            <h2 className={`${styles.headerText} ${styles.display_inline}`}>
+            <h2
+              className={`${styles.headerText} ${styles.display_inline} ${styles.padd_top}`}
+            >
               Manage Databases
             </h2>
             {/* <Button color="yellow" size="md" className={styles.add_mar_left}>
